SGIVideoSyncVsyncMonitorHelper: Unset current context before deletion
We have a driver crash destroying the context / active drawables.

Lets be explicit as possible in ordering so that no other contexts are
potentially active during this time in the hope that it helps work
around the driver problem.

It's not like we'll have to maintain this code for long.

BUG: 507677
(cherry picked from commit a41e49b127e9a950ec1272c808ef04c5e74593fd)